Complete docs, fix tests
