+++ framework-docs/modules/ROOT/pages/testing/mockmvc/playwright.adoc
@@ -1,4 +1,4 @@
-[[mockmvc-server-playwright]]
+[[playwright]]
 = Playwright Integration
 :page-section-summary-toc: 1
 
@@ -7,19 +7,3 @@ https://playwright.dev/[Playwright] via `MockMvcPlaywrightHandler`.
 This allows browser-style end-to-end testing against your MVC controllers without
 starting a server.
 
-== Setup
-
-First, make sure that you have included a test dependency on `com.microsoft.playwright:playwright`.
-To use it, create a`MockMvcPlaywrightHandler` instance and register it as a
-route https://playwright.dev/java/docs/network#handle-requests[handler] for http requests.
-
-include-code::./MockMvcPlaywrightHandlerTests[tag=init,indent=0]
-
-include-code::./MockMvcPlaywrightHandlerTests[tag=setup,indent=0]
-
-Browser http requests, such as form submission and file uploads, are then translated
-to MockMvc requests.
-
-Don't forget to close playwright resources after your tests
-
-include-code::./MockMvcPlaywrightHandlerTests[tag=close,indent=0]
