+++ framework-docs/src/main/java/org/springframework/docs/testing/mockmvc/playwright/mockmvcplaywrighttestingfileupload/MockMvcPlaywrightHandlerTests.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2002-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.docs.testing.mockmvc.playwright.mockmvcplaywrighttestingfileupload;
+
+import com.microsoft.playwright.Page;
+import com.microsoft.playwright.options.FilePayload;
+import org.junit.jupiter.api.*;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit.jupiter.web.SpringJUnitWebConfig;
+import org.springframework.util.StreamUtils;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+
+import static com.microsoft.playwright.assertions.PlaywrightAssertions.assertThat;
+
+@SpringJUnitWebConfig
+public class MockMvcPlaywrightHandlerTests {
+	private Page page;
+
+	// tag::test[]
+	@Test
+	public void testFileUpload() throws IOException {
+		var file = new FilePayload("playwright-single.txt", MediaType.TEXT_PLAIN_VALUE, "single-content".getBytes(StandardCharsets.UTF_8));
+		page.locator("#singleFile").setInputFiles(file);
+		page.locator("#singleFileForm button[type='submit']").click();
+
+		assertThat(page.locator("body")).hasText("file=playwright-single.txt,content=single-content");
+	}
+	// end::test[]
+
+	// tag::controller[]
+	@PostMapping(path = "/upload-single", produces = MediaType.TEXT_PLAIN_VALUE)
+	String uploadSingle(@RequestParam("file") MultipartFile file) throws IOException {
+		return "file=" + file.getOriginalFilename() + ",content=" +
+				StreamUtils.copyToString(file.getInputStream(), StandardCharsets.UTF_8);
+	}
+	// end::controller[]
+}
